requirements.txt
cognitive_service_vision_model_customization_python_samples==0.0.6
GitPython==3.1.43
GitPython==3.1.43
imageio==2.35.1
matplotlib==3.9.2
opencv_python_headless==4.10.0.84
pandas==2.2.3
pdf2image==1.17.0
Pillow==10.0.1
scikit_learn==1.5.2
segment_anything_py==1.0.1
Shapely==2.0.6
scikit-image==0.20.0
streamlit==1.38.0
streamlit_cropper==0.2.2
streamlit_drawable_canvas==0.9.3
tensorflow==2.17.0
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
streamlit-javascript==0.1.5
